requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.8
redis>=3.5
celery>=5.0
django-celery-results>=2.0
django-redis>=5.0
gunicorn>=20.0
djangorestframework 
drf-yasg
requests
pillow
cryptography
python-telegram-bot
django-extensions
phpserialize
django-cors-headers
django-grappelli
python-telegram-bot==20.6